refactor(repo): server modules to server/, previews to previews/, plans to wiki/ [L]
The root held ~25 loose server runtime modules, five dev-preview HTML
harnesses, and orphaned planning docs. Now: server/ holds every server
runtime module (Dockerfile copies it wholesale — the explicit per-file
COPY list and its silent-drop ERR_MODULE_NOT_FOUND trap are retired);
previews/ holds the dev-only render harnesses; the planning docs moved
into wiki/. Boundary paths updated (dist + migrations via __dirname/..,
../scripts imports, node server/server.js in npm start + CI + smoke,
client aiModels import). Lint, build, tests + smoke green; live boot
from the new layout verified serving /api/health and the app shell.
