+++ .github/workflows/build-and-publish.yml
@@ -67,7 +67,7 @@ jobs:
 
       - name: Smoke test
         run: |
-          DB_PATH=./test-smoke.db node server.js &
+          DB_PATH=./test-smoke.db node server/server.js &
           sleep 2
           curl -sf http://localhost:3001/api/health
           curl -sf http://localhost:3001/ | grep -q '<div id="root">'
