+++ CLAUDE.md
@@ -40,13 +40,7 @@
 5. **Run `npm audit` before opening a PR.** Flag new vulnerabilities; fix what's safe (overrides for transitive deps). Don't block on build-time-only vulnerabilities unless asked.
 6. **Build triggers:** a push to `dev` builds `:dev` → `boomerang-dev:3002`; a push to `main` builds `:latest` → prod via Portainer. This is why merge approval matters.
 7. **Proxy realities:** direct `git push origin dev` returns HTTP 403 from the local proxy ("Unable to parse branch information from push data" — fresh-ref pushes and direct `git push origin main` both work; ref deletions 403). So updates to `dev` go through the fresh-ref → PR → merge path below.
-8. **Before EVERY commit that adds or renames files, verify the Docker build picks them up.** The production `Dockerfile` uses an explicit `COPY` list, not `COPY . .`, so new files at the repo root are silently dropped from the container unless you add them. For every new or renamed file in your staged changes:
-   - **Root-level `.js` file imported at runtime** (e.g. `notionMCP.js`, `pushoverNotifications.js`, anything imported by `server.js` or another runtime file): must appear in a Dockerfile `COPY … ./` line in Stage 3. If missing, add it and re-stage. The pre-push smoke test runs `node server.js` against the full repo checkout, so it will NOT catch a missing-from-Dockerfile bug — the container will crash with `ERR_MODULE_NOT_FOUND` only after deploy.
-   - **New top-level directory** that must ship to prod: add a `COPY <dir> ./<dir>` line.
-   - **New migration** under `migrations/`: already covered by the existing `COPY migrations ./migrations` line — no Dockerfile change needed.
-   - **New script** under `scripts/`: already covered by `COPY scripts ./scripts` — no Dockerfile change needed.
-   - **New `src/*` file**: ships via the Vite bundle into `dist/` — no Dockerfile change needed.
-   - **Dev-only files** (`eslint.config.js`, `vite.config.js`, tests, docs): do NOT add to the runtime Dockerfile.
+8. **Server runtime modules live in `server/` (2026-07-16 reorg) — new server code goes THERE, and the Dockerfile copies the directory wholesale (`COPY server ./server`).** This retired the old explicit per-file COPY list and its silent-drop trap (a new root module missing from the list crashed prod with `ERR_MODULE_NOT_FOUND` only after deploy). What still needs a Dockerfile edit: a genuinely NEW top-level directory that must ship to prod (`COPY <dir> ./<dir>`). Already covered, no edit needed: anything in `server/`, `migrations/`, `scripts/`, and all `src/*` (ships via the Vite bundle into `dist/`). Dev-only files (`eslint.config.js`, `vite.config.js`, `previews/`, tests, docs) stay OUT of the image. Path conventions inside `server/`: modules import each other as `./x.js`; reaching repo-root resources uses `..` (`../scripts/backup-db.js`, `path.join(__dirname, '..', 'dist')`, `path.join(__dirname, '..', 'migrations')`); the shared `server/aiModels.js` is imported by the client as `../server/aiModels.js` from `src/`. NOTE: older sections of this file say "root module, in the Dockerfile runtime COPY list" — historical phrasing from before the reorg; those modules are all in `server/` now.
 
 ### Workflow: develop on `dev`
