+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "start": "node server.js",
+    "start": "node server/server.js",
     "lint": "eslint .",
     "test": "node --test scripts/dates.test.mjs scripts/cycles.test.mjs scripts/impact.test.mjs && sh scripts/smoke-test.sh",
     "prepare": "git config core.hooksPath .githooks || true",
