+++ server/adviserTools.js
