+++ server/adviserToolsMisc.js
