+++ server/digestBuilder.js
