+++ server/knowledgeSync.js
