+++ server/patternDetection.js
