+++ server/weatherSync.js
